examples/backends/fastapi/pyproject.toml
[project]
name = "l42-token-handler-fastapi"
version = "0.8.0"
description = "FastAPI Token Handler backend for l42-cognito-passkey"
requires-python = ">=3.11"
license = "Apache-2.0"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "pyjwt[crypto]>=2.10.0",
    "cedarpy>=4.8.0",
    "itsdangerous>=2.2.0",
    "aioboto3>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "cryptography>=44.0.0",
    "pytest-cov>=6.0.0",
    "moto[dynamodb]>=5.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["cedar*", "tests*"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
